_legacy/pipeline/pipeline/requirements.txt
anthropic>=0.42.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
praw>=7.8.0
python-dotenv>=1.0.0
